Add module "likes.js" (Follow ES6 good practices)
