+++ src/modules/likes.js
@@ -1,0 +1,24 @@
+const InvAPIurl = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/aZwedqY3IXDWCJKZoRkM/likes';
+
+const getLikes = async () => {
+  try {
+    const response = await fetch(InvAPIurl, {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
+    if (!response.ok) {
+      throw new Error('Network response was not ok');
+    }
+    const contentType = response.headers.get('content-type');
+    if (contentType && contentType.includes('application/json')) {
+      return response.json();
+    }
+    return []; // Return an empty array if the response body is empty or not in JSON format
+  } catch (error) {
+    return []; // Return an empty array in case of an error
+  }
+};
+
+export default getLikes;
