+++ webpack.config.js
@@ -9,6 +9,7 @@ module.exports = {
     renderHeaderFooter: './src/modules/render_header_footer.js',
     popup: './src/modules/popup.js',
     home: './src/modules/home_page.js',
+    likes: './src/modules/likes.js',
   },
   devtool: 'inline-source-map',
   devServer: {
